separate logPage in components
