+++ prisma/schema.prisma
@@ -13,6 +13,7 @@ model User {
   id    Int     @id @default(autoincrement())
   email String  @unique
   name  String?
+  password String
   // posts Post[]
 }
